1531/Group project/project-backend/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
PyJWT==2.0.1
requests==2.25.1


